requirements.txt
streamlit
langchain-community
python-dotenv
langchain
langchain-groq
pandas
langchain-google-genai
PyPDF2